feat: Update `DirectMessagePopup` component
